FINAL PUSH

+++ includes/Server.hpp
@@ -61,7 +61,7 @@
             string                      _root;
             string                      _storage_dir;
             LocationsList               _locations;
-            int                         _request_body_size_limit;
+            long                         _request_body_size_limit;
             string                      _index;
             string                      _server_name;
 
@@ -99,7 +99,7 @@
             string          get_storage_dir() const { return _storage_dir; }
             LocationsList   get_locations() { return _locations; }
             string          get_root() { return _root; }
-            int             get_max_body_size() { return _request_body_size_limit; }
+            long             get_max_body_size() { return _request_body_size_limit; }
 
             LocationsList::iterator lbegin();
             LocationsList::iterator lend();
